feat(core): shared types, SDK, operators, and extractor improvements
- Update shared types for pipeline, step, trigger, adapter, and execution
- Refine SDK DSL helpers and validation utilities
- Improve operator helpers for aggregation, enrichment, and JSON operations
- Update extractor schemas for GraphQL and HTTP API
- Enhance parsers, transforms, and feed generators
- Update entity definitions and constant enumerations

+++ src/entities/pipeline/pipeline-log.entity.ts
@@ -46,7 +46,7 @@ export class PipelineLog extends VendureEntity {
     @Column({ type: 'int', nullable: true })
     runId!: number | null;
 
-    @Column({ type: 'bigint', nullable: true })
+    @Column({ type: 'int', nullable: true })
     durationMs!: number | null;
 
     @Column({ type: 'int', nullable: true })
